+++ monai/networks/blocks/selfattention.py
@@ -15,8 +15,8 @@ from typing import Optional, Tuple
 
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
+from monai.networks.layers.utils import get_rel_pos_embedding_layer
 from monai.utils import optional_import
 
 Rearrange, _ = optional_import("einops.layers.torch", name="Rearrange")
@@ -36,7 +36,7 @@ class SABlock(nn.Module):
         dropout_rate: float = 0.0,
         qkv_bias: bool = False,
         save_attn: bool = False,
-        use_rel_pos: Optional[str] = None,
+        rel_pos_embedding: Optional[str] = None,
         input_size: Optional[Tuple] = None,
     ) -> None:
         """
@@ -45,7 +45,7 @@ class SABlock(nn.Module):
             num_heads (int): number of attention heads.
             dropout_rate (float, optional): fraction of the input units to drop. Defaults to 0.0.
             qkv_bias (bool, optional): bias term for the qkv linear layer. Defaults to False.
-            rel_pos (str, optional): Add relative positional embeddings to the attention map.
+            rel_pos_embedding (str, optional): Add relative positional embeddings to the attention map.
                 For now only "decomposed" is supported (see https://arxiv.org/abs/2112.01526). 2D and 3D are supported.
             input_size (tuple(spatial_dim), optional): Input resolution for calculating the relative
                 positional parameter size.
@@ -72,14 +72,12 @@ class SABlock(nn.Module):
         self.scale = self.head_dim**-0.5
         self.save_attn = save_attn
         self.att_mat = torch.Tensor()
-        self.use_rel_pos = use_rel_pos
+        self.rel_positional_embedding = (
+            get_rel_pos_embedding_layer(rel_pos_embedding, input_size, self.head_dim, self.num_heads)
+            if rel_pos_embedding is not None
+            else None
+        )
         self.input_size = input_size
-
-        if self.use_rel_pos == "decomposed":
-            assert input_size is not None, "Input size must be provided if using relative positional encoding."
-            self.rel_pos_arr = nn.ParameterList(
-                [nn.Parameter(torch.zeros(2 * dim_input_size - 1, self.head_dim)) for dim_input_size in input_size]
-            )
 
     def forward(self, x: torch.Tensor):
         """
@@ -93,18 +91,8 @@ class SABlock(nn.Module):
         q, k, v = output[0], output[1], output[2]
         att_mat = torch.einsum("blxd,blyd->blxy", q, k) * self.scale
 
-        if self.use_rel_pos == "decomposed":
-            batch = x.shape[0]
-            h, w = self.input_size[:2] if self.input_size is not None else (0, 0)
-            d = self.input_size[2] if self.input_size is not None and len(self.input_size) > 2 else 1
-            att_mat = add_decomposed_rel_pos(
-                att_mat.view(batch * self.num_heads, h * w * d, h * w * d),
-                q.view(batch * self.num_heads, h * w * d, -1),
-                self.rel_pos_arr,
-                (h, w) if d == 1 else (h, w, d),
-                (h, w) if d == 1 else (h, w, d),
-            )
-            att_mat = att_mat.reshape(batch, self.num_heads, h * w * d, h * w * d)
+        # apply relative positional embedding if defined
+        att_mat = self.rel_positional_embedding(x, att_mat, q) if self.rel_positional_embedding is not None else att_mat
 
         att_mat = att_mat.softmax(dim=-1)
 
@@ -119,88 +107,3 @@ class SABlock(nn.Module):
         x = self.out_proj(x)
         x = self.drop_output(x)
         return x
-
-
-def get_rel_pos(q_size: int, k_size: int, rel_pos: torch.Tensor) -> torch.Tensor:
-    """
-    Get relative positional embeddings according to the relative positions of
-        query and key sizes.
-    Args:
-        q_size (int): size of query q.
-        k_size (int): size of key k.
-        rel_pos (Tensor): relative position embeddings (L, C).
-
-    Returns:
-        Extracted positional embeddings according to relative positions.
-    """
-    rel_pos_resized: torch.Tensor = torch.Tensor()
-    max_rel_dist = int(2 * max(q_size, k_size) - 1)
-    # Interpolate rel pos if needed.
-    if rel_pos.shape[0] != max_rel_dist:
-        # Interpolate rel pos.
-        rel_pos_resized = F.interpolate(
-            rel_pos.reshape(1, rel_pos.shape[0], -1).permute(0, 2, 1), size=max_rel_dist, mode="linear"
-        )
-        rel_pos_resized = rel_pos_resized.reshape(-1, max_rel_dist).permute(1, 0)
-    else:
-        rel_pos_resized = rel_pos
-
-    # Scale the coords with short length if shapes for q and k are different.
-    q_coords = torch.arange(q_size)[:, None] * max(k_size / q_size, 1.0)
-    k_coords = torch.arange(k_size)[None, :] * max(q_size / k_size, 1.0)
-    relative_coords = (q_coords - k_coords) + (k_size - 1) * max(q_size / k_size, 1.0)
-
-    return rel_pos_resized[relative_coords.long()]
-
-
-def add_decomposed_rel_pos(
-    attn: torch.Tensor, q: torch.Tensor, rel_pos_lst: nn.ParameterList, q_size: Tuple, k_size: Tuple
-) -> torch.Tensor:
-    """
-    Calculate decomposed Relative Positional Embeddings from :paper:`mvitv2`.
-    https://github.com/facebookresearch/mvit/blob/19786631e330df9f3622e5402b4a419a263a2c80/mvit/models/attention.py   # noqa B950
-    Only 2D and 3D are supported.
-    Args:
-        attn (Tensor): attention map.
-        q (Tensor): query q in the attention layer with shape (B, s_dim_1 * ... * s_dim_n, C).
-        rel_pos_lst (ParameterList): relative position embeddings for each axis: rel_pos_lst[n] for nth axis.
-        q_size (Tuple): spatial sequence size of query q with (q_dim_1, ..., q_dim_n).
-        k_size (Tuple): spatial sequence size of key k with (k_dim_1, ...,  k_dim_n).
-
-    Returns:
-        attn (Tensor): attention map with added relative positional embeddings.
-    """
-    rh = get_rel_pos(q_size[0], k_size[0], rel_pos_lst[0])
-    rw = get_rel_pos(q_size[1], k_size[1], rel_pos_lst[1])
-
-    batch, _, dim = q.shape
-
-    if len(rel_pos_lst) == 2:
-        q_h, q_w = q_size[:2]
-        k_h, k_w = k_size[:2]
-        r_q = q.reshape(batch, q_h, q_w, dim)
-        rel_h = torch.einsum("bhwc,hkc->bhwk", r_q, rh)
-        rel_w = torch.einsum("bhwc,wkc->bhwk", r_q, rw)
-
-        attn = (attn.view(batch, q_h, q_w, k_h, k_w) + rel_h[:, :, :, :, None] + rel_w[:, :, :, None, :]).view(
-            batch, q_h * q_w, k_h * k_w
-        )
-    elif len(rel_pos_lst) == 3:
-        q_h, q_w, q_d = q_size[:3]
-        k_h, k_w, k_d = k_size[:3]
-
-        rd = get_rel_pos(q_d, k_d, rel_pos_lst[2])
-
-        r_q = q.reshape(batch, q_h, q_w, q_d, dim)
-        rel_h = torch.einsum("bhwdc,hkc->bhwdk", r_q, rh)
-        rel_w = torch.einsum("bhwdc,wkc->bhwdk", r_q, rw)
-        rel_d = torch.einsum("bhwdc,wkc->bhwdk", r_q, rd)
-
-        attn = (
-            attn.view(batch, q_h, q_w, q_d, k_h, k_w, k_d)
-            + rel_h[:, :, :, :, None, None]
-            + rel_w[:, :, :, None, :, None]
-            + rel_d[:, :, :, None, None, :]
-        ).view(batch, q_h * q_w * q_d, k_h * k_w * k_d)
-
-    return attn
